Improves resilience to broken stdout/stderr
Addresses potential crashes caused by `EPIPE` or `EIO` errors when `stdout` or `stderr` are disconnected, such as when a watchdog process kills the main process.

Instead of crashing when logging fails due to a broken pipe, the changes implement a `safeLog` function to gracefully handle write errors to `console.log`. Error logging is also updated to check for both `EPIPE` and `EIO` errors, logging to a file instead of the console when these occur.